Add CSS for payment section
Created a new file _payment.scss which contains CSS for the payment section. This styling includes responsive properties for different screen sizes, animations, highlights, platform text, and specialized styles for the payment tooltip and modals. This ensures a better user interface and enhances the overall aesthetics of the payment section.