Change hero area background image and remove unnecessary import
In `hero-area.jsx`, the background image was changed to `hero-bg-2.png` in order to update the visual design of the homepage. In `header-2.jsx`, an unnecessary import statement was removed for `white_logo`, as it was declared twice. This change will help to make the code cleaner and more efficient.