Change CTA link to direct to registration page
Updated the call to action (CTA) button's link in the home component to direct users to the registration page instead of the service details page. This change makes it more convenient for new users to register for the service right from the homepage, thus improving user experience and potentially increasing the number of sign-ups.